rpmRecordParser::LongDesc(): restore the old format (expected by clients) (ALT#40826)
This reverts a part of commit 1ebe1d7079b5f12f0fa145e18ed60eb30660ad2e.

This makes the format again closer to Debian's descriptions[1],
which of course are mainly expected by all the libapt clients like
aptitude, maybe synaptic, packagekit etc., including any new code there,
which has not been adapted to apt-rpm.

[1]: https://www.debian.org/doc/debian-policy/ch-controlfields.html#description

The fixed discrepancy with Debian: all but the first lines begin with a space.
So, the parsers in the libapt clients usually delete the first character--
no matter whether it actually is a space or not.
How clients are affected:
* aptitude: subsequent lines not displayed ALT#40826;
* packagekit: first characters not displayed by "pkcon get-details"; also,
for some reason, this change fixes the problem that Russian translations
were shown as empty ALT#46251 (perhaps, that's due to the encoding);
* synaptic: not really (because it uses "no parser" by default for rpm);
and after this fix, it shows an extra space in the beginnings of lines (FIXME).

A non-fixed discrepancy with Debian: the first line contains the short summary.
So, the parsers in the libapt clients usually delete the first line.
How clients are affected:
* packagekit: does delete the first line (FIXME);
* synaptic: not affected because it is adapted to apt-rpm by using "no parser";
* aptitude: not affected because it's evidently adapted to apt-rpm

A non-fixed discrepancy with Debian: a line containing a single space followed
by a single full stop character is the only way to get a blank line.
So, a pretty-printer may dump all other lines into a single paragraph and wrap
the whole paragraph to its own liking (in a way suitable for the UI).
How clients are affected:
* synaptic: no, it preserves original formatting as long as it uses "no parser";
* aptitude: dumps all lines and words into a single paragraph, even separated
by blank lines in the original RPM description;
* packagekit: same small problem as in aptitude. (FIXME)

Another (unused) feature of the Debian format: lines starting with two
or more spaces are to be displayed verbatim.
We currently don't try to mark up this way lines from the RPM descriptions
that shouldn't be dumped into a paragraph (like list items etc.) (FIXME,
but no good way to guess.)