rpmRecordParser::LongDesc(): insert extra 1st blank line (closer to Debian)
This mitigates the problem that the parsers in some libapt clients
expect an extra first line (like in Debian) and delete it.

The partly fixed discrepancy with Debian: the first line contains the short summary.
So, the parsers in the libapt clients usually delete the first line.
How clients are affected:
* packagekit: does delete the first line (fixed by this change);
* synaptic: was not affected because it is adapted to apt-rpm by using
"no parser"; with this change, there'll be a blank line, not too bad.
(Later, synaptic can be made use the "deb parser" by a config var.)
* aptitude: was not affected because it's evidently adapted to apt-rpm.
(Later, aptitude's code can be reverted to the old Debian behavior.)