configure: do the fallback onto the same archname
...instead of having to extend the archtable file every time there is
a new arch.

Do it in configure instead of .spec, so that the build procedure works
the same (and as expected) even when built not under rpm.

The trivial translations (to the same name) can be removed from the
archtable file (whose only purpose BTW was for the build procedure,
it's not used at run-time). But we can leave them so that configure.ac
warns us if it isn't able to find an actually known arch name
(perhaps, due to a bug in the awk script, which we are going to
simplify).