Implementation of 'apt-get changelog'
The code is based on ALT's 'apt-cache show' (see cmdline/apt-cache.cc).
Virtual packages handling is taken from 'show' implementation.
Utility takes changelog from Candidate version (i.e. the most recent).
'rpmRecordParser' is used to extract Changelog.

Usage example:
 ```
    $ apt-get changelog gcc htop
 ```
 Will display changelog of 'gcc' and 'htop' packages

[imz: Logically, in ALT's apt-rpm, unlike in Debian, the "changelog"
command belongs to apt-cache, since it doesn't need to get anything,
just read the caches. It's in apt-get for cmdline compatibility.
TODO: add it to apt-shell and apt-cache.

DoChangelog() may be desired to make output similar to "apt-cache show"
in that that it shouldn't show the progress of building the caches. However:

Currently, passing devnull in "CacheFile Cache(devnull, ...);"
would affect only the output of CacheFile::CheckDeps(), which is not
employed in DoChangelog(). As for Cache.Open(), its output of the
progress of building the caches is not affected, i.e., it is shown anyway.
This makes this piece of output different from how it would be in apt-cache.
]