tests: an option to write built pkgs under the filename of our alias for the pkg
We put %{ARCH}/ as the prefix, because some other code in our tests wants
to find the files under such subdirs.

We need to put at least ./ (or %{ARCH}/) in the filename in order to
make rpmbuild create the dir. The relevant code from rpmbuild, which
does MkdirP only if there is a dir component (pack.c):

	    if ((binDir = strchr(binRpm, '/')) != NULL) {
		struct stat st;
		const char *dn;
		*binDir = '\0';
		dn = rpmGetPath("%{_rpmdir}/", binRpm, NULL);
		if (Stat(dn, &st) < 0) {
		    switch(errno) {
		    case  ENOENT:
			if (MkdirP(dn, 0755) == 0)
			    /*@switchbreak@*/ break;
			/*@fallthrough@*/