Fix ShowEssential output
Normally, packages in List are separated by a space; but when we
are called from ShowEssential, List has additional information,
and additional spaces. In this case, List cannot be reliably
split into packages, so we should not use columnar output. This
case can be detected by checking if VersionsList is empty.

This seems to be the least invasive fix for ALT#51546