tests: test cert pinning each time aptget is called in any tests (https_pinned method)
Now, no cert is pinned in tests with the usual "https" method anymore,
but pinning is always tested with the new "https_pinned" quasi-method.

This looks like overdoing the testing of one feature (the access to
the repo with verifying the pinned cert), but well, it doesn't
hurt: who knows in which circumstances something works unexpectedly.

And I'm going to use the same approach in future for testing the
access to the repo through proxy (not directly). So, it's a
preparation of the scheme.

* * *

The approach still looks ugly because we are abusing other tests (with
their own intentions) for covertly testing another feature (of a
download method). Well, we can look at this not like that an
individual test gets another purpose, but that the whole set of tests
is a test for a download method to function correctly.