test-...-zillion: no random names; avoid collisions by locking
We allow concurrent re-use of the generated specs by locking the same
way it is done in buildpackage().

Also, we've adapted buildpackage() to find specs under INTERMEDIATES,
too, so we can save our generated specs there.

Basically, the locking and re-use of specs is not for spped (that's fast),
but to avoid collisions when parallel processes do the same.
(And that was the reason for the old random names.)

As for speed, we still might get an improvement due to buildpackage() being
able to re-use built packages thanks to persistent names.