test-apt-method-https-verifies-pinned-cert: also check the valid cert (among https tests)
Now, that cert pinning is no longer done during the https tests
(whereas it's done for a separate https_pinned pseudo-method),
it makes sense to check (during the https tests) that pinning with the
valid cert works, too.

So, we make this test more complete again: it checks both invalid and
valid certs; the logic is completely independent from the https_pinned
pseudo-method (which itself might be considered an overkill, but at
least it is a proof-of-concept for running all the tests with a
special configuration and having more probability of catching
something strange with the special configuration in one of the cases.
This idea of mass-testing really helped later to discover a connection
bug when running all the tests with proxy, even though the bug was not
directly related to proxy code being broken--just more special
conditions allow to expose more bugs.)