test*(): don't wrap the run cmds with "set +e/-e" (misbehaved when test*() nest)
Instead, wrap them with "|| result=$?".

Now we've seen that a nested testfailure() would do "set -e" and therefore
break the non-aborting mode of the command given to the top-level testfailure().

FIXME:
Another thing that needs to be fixed for nested testfailure()'s (etc.)
to work properly is the collision on $OUTPUT. (Perhaps mktemp?)
So, for now, avoid nesting.