tests: implement http*_proxy* quasi-methods with the use of tinyproxy
And we do XFAIL when HTTPS through proxy is actually done in a test:
We have yet to implement this feature.
And now we just have the tests ready for it.

We prioritize checking pinning over ensuring proxy-only access. So we
override aptget() for checking pinning after the http*_proxy*
override.

* * *

Running all the tests with one more quasi-method just to test one feature
instead of having a single test might seem as an overkill. But let's do
it for the following benefits:

1. this way, we massively test apt (not only the proxy feature) under
more different conditions; and this can uncover an unnoticed bug in apt
in a place where we don't expect.

2. we ensure that the feature always works.

3. we make the test framework and the set of tests more robust, ready to be
adapted to different conditions preserving the meaning of the tests.

Reason 1 actually played out this time: we saw a few of failed downloads,
thanks to very special conditions created in some tests just by coincidence,
and this exposed a bug in apt.