APT::Get::List-Columns "false" by default for now because of a shortcoming
Since in the code we re-parse a plain string into a list by
splitting it into words (unlike Debian, where the code gets a list
structure internally), now we saw some troubles[1] when the items had
spaces (like those with a space and parentheses). Therefore, it's off
by default for now; and it's also an easy way to build APT for p10
with some fixes without changing its appearance. (This commit can be
reverted later in Sisyphus, or not applied at all, but used for
stable branches.)

[1]: https://bugzilla.altlinux.org/51546