tests(nginxsetup): no unset local vars, to spot usage by wrong names
If we make a typo in the name of var when it is used, but generally we
expect that the right name can be unset, it won't be detected as an
error. So we should try to write the code (where reasonable) in the
assumption that the vars have to be set.