test-apt-method-http*-uses-*proxy*: run only if proxy is configured (to simplify reconfiguration)
In order not to have the complexity to reconfigure tinyproxy
(tinyproxysetup with the upcoming various options for IPv?, numeric),
not to bring the comlexity into these scripts, just run them if
the quasi-method has proxy configured.

So, the purpose of these test scripts is to make sure that there is no
access without proxy and hence be sure that proxy is actually used on
the final step with various ways to give such configuration to apt.