test-*-verifies-pinned-cert: reshape per the kind of common scheme: re-use the method setup
As in other tests of this kind now, we let the common setup code to
set up the repo and servers and apt config for a specific quasi-method
(in this case: with pinning), and then the scheme is to run apt with
the crucial part of the config undone, make sure apt fails, and then
see that it works with the config restored: this is a proof that this
part of config actually affects apt in a way we wish.

As a future re-organization in a move to reduce the number of tests
run and the excessive meaningless method modifiers, we can make such
tests modify the current method with a modifier corresponding to the
tested feature and do such a setup just for this test, without
meaninglessly running a lot of tests with such modified method.

For instance, it's similar to:

diff test-apt-method-http-connects-to-nonlocalhost test-apt-method-https-verifies-pinned-cert