run-tests: restrict some quasi-methods to http tests (to run less testcases)
The quasi-methods removed from the general list do not affect the
connection after it is established or the download paths. So, there is
no reason to apply them to all the other tests except those specific
ones that check whther a connection is established. (As for the
*_proxy quasi-methods: they remain, because things can be sent/read
differently through the proxy, and this has uncovered some bugs in the
http code.)