protect Debug::Connect output with S_IRUSR mode on the files (just in case...)
In case (lib)apt runs with different privileges than the calling user,
we have a question: what additional risks does this Debug::Connect
feature bring?

1. Overwriting a file which is not accessible to the unprivileged user.
2. Exposing secret information which is not known to the user.

It's safer not to assume that in such case (if it happens to be run
set-UID) the user wouldn't be able to set the config option and not to
assume that there is no secret information in the connection internals
(think about auth), but to protect from both risks: with O_EXCL and
S_IRUSR on the output files.

Amends: fbb5dec21 (debugging of MethodFd by wrapping, 2025-03-20)