https subpkg: drop it (a separate entity doesn't make sense after joining)
The only thing that I see that could hold us back from committing to
joining http and https (as was done in the previous commit) is that
we'd additionally get gnutls in a basic installation.

What other basic packages are there now that depend on libgnutls30? I
see: gnupg2, libcurl. SO maybe it's not a big deal.

Debian's apt has switched to openssl. systemd, too. The reason is
probably the understanding that gnutls code is not nice and clean and
has a history of bugs.