methods/Makefile.am: better target for symlinks (per automake docs)
It's more semantically appropriate to put this into a *-hook, because
we count on the base methods to be installed. (This is however also
expressed with an order-only prerequisite). From the automake docs:

With the -local targets, there is no particular guarantee of execution
order; typically, they are run early, but with parallel make, there is
no way to be sure of that.

In contrast, some rules also have a way to run another rule, called a
hook; hooks are always executed after the main rule’s work is done.