%%changelog: amend 0.5.15lorg2-alt90 with an OVE id
OVE-20250602-0001: Failure to complete an HTTP download may prevent a timely security update.

An error in APT's http method may prevent it to complete an HTTP download and to do a security update on time.

(In practice, this was observed intermittently when working through
a specific proxy server, namely tinyproxy, due to specific sizes of
the chunks in which data was transferred.)

Workaround: try more times, or download the file with an external tool
and put it into APT's download cache directory.