%%changelog: amended 0.5.15lorg2-alt96 with an OVE id
OVE-20250602-0002: Not using HTTPS from behind a proxy opens doors for exploiting APT's bugs.

When using APT from behind a proxy, impossibility to use a trusted
connection to the trusted vendor's HTTPS server may open doors for an
attack on APT's http method that would make APT misbehave and execute
arbitrary code (in a package) prepared by an attacker who controls one
of the connection nodes, e.g., by exploiting any bug in APT that would
lead to misinterpretation of internal responses from the http method,
which--among other things--has the power to fake computed checksums.

For an idea/an example of a potential way of an attack
see CVE-2019-3462
https://www.akamai.com/blog/security/vulnerability-in-debians-advanced-package-tool .